requirements.txt
python-dotenv==1.0.0
streamlit==1.28.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langchain-google-genai==1.0.0
openai==1.3.0
pypdf==3.17.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
google-generativeai==0.3.0
